# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Taal API
  description: 'Demand-shaping plays with a holdout: sense, plan, approve, engage,
    measure.'
  version: 0.1.0
paths:
...
  /gaps:
    get:
      summary: Gaps
      operationId: gaps_gaps_get
      parameters:
      - name: node_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Node Id
      - name: type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response Gaps Gaps Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /gaps/{gap_id}:
    get:
      summary: Gap
      operationId: gap_gaps__gap_id__get
      parameters:
      - name: gap_id
        in: path
        required: true
        schema:
          type: string
          title: Gap Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Gap Gaps  Gap Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /plays:
    get:
      summary: Plays
      operationId: plays_plays_get
      parameters:
      - name: gap_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Gap Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response Plays Plays Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /plays/{play_id}:
    get:
      summary: Play
      operationId: play_plays__play_id__get
      parameters:
      - name: play_id
        in: path
        required: true
        schema:
          type: string
          title: Play Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Play Plays  Play Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /plan:
    post:
      summary: Plan
      operationId: plan_plan_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Plan Plan Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /approve:
    post:
      summary: Approve Play
      operationId: approve_play_approve_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Approve Play Approve Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /rerun:
    post:
      summary: Rerun
      operationId: rerun_rerun_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RerunRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Rerun Rerun Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /policy:
    get:
      summary: Policy
      operationId: policy_policy_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Policy Policy Get
    put:
      summary: Put Policy
      operationId: put_policy_policy_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Put Policy Policy Put
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /events/{run_id}:
    get:
      summary: Events
      operationId: events_events__run_id__get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Events Events  Run Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /events/{run_id}/stream:
    get:
      summary: Events Stream
      operationId: events_stream_events__run_id__stream_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: speed
        in: query
        required: false
        schema:
          type: number
          maximum: 100
          minimum: 0.1
          default: 4.0
          title: Speed
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /chat:
    post:
      summary: Chat
      operationId: chat_chat_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /capture:
    post:
      summary: Capture
      operationId: capture_capture_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaptureRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Capture Capture Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /capture/confirm:
    post:
      summary: Capture Confirm
      operationId: capture_confirm_capture_confirm_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaptureConfirmRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Capture Confirm Capture Confirm Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /customers/demo:
    get:
      summary: Customers Demo
      description: 'A curated picker for the web app: named personas with visibly
        different home stores and

        languages, plus one real customer who lands in the holdout arm of `play_id`
        -- computed from

        the play''s own holdout seed and fraction, so it is correct whether or not
        the play has been

        approved yet, and never made up. Lets a judge see the differentiation the
        system is built on:

        a holdout customer''s chat never mentions the offer other customers get.'
      operationId: customers_demo_customers_demo_get
      parameters:
      - name: play_id
        in: query
        required: false
        schema:
          type: string
          default: play_chips_ds07_v1
          title: Play Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response Customers Demo Customers Demo Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /execution:
    post:
      summary: Execution
      operationId: execution_execution_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Execution Execution Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ApproveRequest:
      properties:
        play_id:
          type: string
          title: Play Id
        holdout_fraction:
          anyOf:
          - type: number
            maximum: 0.9
            minimum: 0.05
          - type: 'null'
          title: Holdout Fraction
        rationale:
          anyOf:
          - type: string
          - type: 'null'
          title: Rationale
        edits:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Edits
        approved_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Approved By
      type: object
      required:
      - play_id
      title: ApproveRequest
    CaptureConfirmRequest:
      properties:
        node_id:
          type: string
          title: Node Id
        photo_ref:
          type: string
          title: Photo Ref
        rows:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Rows
      type: object
      required:
      - node_id
      - photo_ref
      - rows
      title: CaptureConfirmRequest
    CaptureRequest:
      properties:
        node_id:
          type: string
          title: Node Id
        photo_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Photo Ref
        image_data_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Data Url
      type: object
      required:
      - node_id
      title: CaptureRequest
    ChatRequest:
      properties:
        session_id:
          type: string
          pattern: ^[A-Za-z0-9_-]+:(web|whatsapp)$
          title: Session Id
        text:
          type: string
          maxLength: 2000
          title: Text
        customer_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Customer Id
        language:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
      type: object
      required:
      - session_id
      - text
      title: ChatRequest
    ExecutionRequest:
      properties:
        play_id:
          type: string
          title: Play Id
        node_id:
          type: string
          title: Node Id
        steps_done:
          items:
            type: string
          type: array
          title: Steps Done
        evidence_photo_data_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Evidence Photo Data Url
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      required:
      - play_id
      - node_id
      - steps_done
      title: ExecutionRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    PlanRequest:
      properties:
        gap_id:
          type: string
          title: Gap Id
      type: object
      required:
      - gap_id
      title: PlanRequest
    PolicyUpdate:
      properties:
        text:
          type: string
          title: Text
        policy_version:
          type: string
          title: Policy Version
      type: object
      required:
      - text
      - policy_version
      title: PolicyUpdate
    RerunRequest:
      properties:
        gap_id:
          type: string
          title: Gap Id
        policy_text:
          type: string
          title: Policy Text
        policy_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Policy Version
      type: object
      required:
      - gap_id
      - policy_text
      title: RerunRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError